refactor(#85): deduplicate ask_user_types option parsing, document headless test split
Extract shared parse_options/parse_option into ask_user_types/mod.rs.
canonical.rs and flat.rs now delegate to super::parse_options — eliminates
~28 lines of duplicated option-parsing logic. The preview field fallback
(previously only in flat.rs) is now shared, which is a benign extension.

Add doc comment to executor/headless.rs explaining why tests live in the
executor file rather than alongside core.rs.
